Update player.h
Added a setter function, chooseClass, that modifies base player stats.  This is distinct from choosing Armor, and other items that affect the total stats but not the base stats.  I tested it on my computer first.

Since this is a very simple member function, it is defined in the header file rather than in adventure.cpp